+++ src/geometrylib/perimeter.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #define _USE_MATH_DEFINES
 #include <stdio.h>
-float calculate_perimeter(float rad)
+double calculate_perimeter(double rad)
 {
     return rad * 2 * M_PI;
 }
